[ResponseOps] Move custom threshold rule params to the package (#208686)
Fixes: https://github.com/elastic/kibana/issues/195191

Move log threshold rule type params to the new package.

P.S.: I've moved function `validateKQLStringFilter` and test for it in
my previous PR: https://github.com/elastic/kibana/pull/205507

---------

(cherry picked from commit 07557b686c79c2af9dd3b2789c66efc1f6ca840d)
