+++ x-pack/solutions/observability/plugins/observability/server/lib/rules/custom_threshold/utils.test.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { flattenObject, getFormattedGroupBy, validateKQLStringFilter } from './utils';
+import { flattenObject, getFormattedGroupBy } from './utils';
 
 describe('FlattenObject', () => {
   it('flattens multi level item', () => {
@@ -54,29 +54,6 @@ describe('FlattenObject', () => {
       'key1.item1': 'value 1',
       'key1.item2.itemA': 'value 2',
     });
-  });
-});
-
-describe('validateKQLStringFilter', () => {
-  const data = [
-    // input, output
-    ['', undefined],
-    ['host.name:host-0', undefined],
-  ];
-  const dataWithError = [
-    // input, output
-    [
-      ':*',
-      'filterQuery must be a valid KQL filter (error: Expected "(", NOT, end of input, field name, value, whitespace but ":" found.',
-    ],
-  ];
-
-  test.each(data)('validateKQLStringFilter(%s): %o', (input: any, output: any) => {
-    expect(validateKQLStringFilter(input)).toEqual(output);
-  });
-
-  test.each(dataWithError)('validateKQLStringFilter(%s): %o', (input: any, output: any) => {
-    expect(validateKQLStringFilter(input)).toContain(output);
   });
 });
