+++ x-pack/solutions/observability/plugins/observability/tsconfig.json
@@ -114,7 +114,8 @@
     "@kbn/response-ops-rule-form",
     "@kbn/streams-plugin",
     "@kbn/data-service",
-    "@kbn/ebt-tools"
+    "@kbn/ebt-tools",
+    "@kbn/response-ops-rule-params"
   ],
   "exclude": ["target/**/*"]
 }
